backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
gitpython==3.1.40
chromadb==0.4.22
sentence-transformers==2.3.1
groq==0.11.0
redis==5.0.1
pydantic==2.5.3